pyproject.toml
[project]
name = "onedrive-photo-namer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "ffmpeg-python>=0.2.0",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pytz>=2025.2",
]
